2683335 add cast to avoid compile warning by arm StartUp_Arch.c
git-svn-id: https://svn.code.sf.net/p/opensek/code/trunk@213 73ee5bea-b2a1-4b4e-8992-c3b1447581ab

+++ OpenSEK/src/arm/StartOs_Arch.c
@@ -69,7 +69,7 @@ void StartOs_Arch(void)
 		TasksConst[loopi].TaskContext->reg_r13 = (uint32)TasksConst[loopi].StackPtr + TasksConst[loopi].StackSize;
 
 		/* init entry point */
-		TasksConst[loopi].TaskContext->reg_r15 = TasksConst[loopi].EntryPoint;
+		TasksConst[loopi].TaskContext->reg_r15 = (uint32) TasksConst[loopi].EntryPoint;
 
 		/* init program status register */
 		TasksConst[loopi].TaskContext->reg_cpsr = 0x000000d3; /** ?? TODO */
